+++ src/android/com/totalpave/cordova/insets/Insets.java
@@ -16,6 +16,7 @@
 
 package com.totalpave.cordova.insets;
 
+import android.content.Context;
 import android.os.Build;
 import android.view.RoundedCorner;
 import android.view.WindowInsets;
@@ -29,16 +30,15 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.JSONException;
 import java.lang.NumberFormatException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.UUID;
 
 public class Insets extends CordovaPlugin {
     public static final int DEFAULT_INSET_MASK = WindowInsetsCompat.Type.displayCutout() | WindowInsetsCompat.Type.systemBars();
+    public static final boolean DEFAULT_INCLUDE_ROUNDED_CORNERS = true;
 
-    public CallbackContext listener;
-    private JSONObject insets;
-    private int insetMask;
-    
-
-    private static class WebviewMask {
+    public static class WebviewMask {
         private WebviewMask() {}
 
         public static final int CAPTION_BAR                 = 1;
@@ -52,28 +52,58 @@ public class Insets extends CordovaPlugin {
         public static final int TAPPABLE_ELEMENT            = 1 << 8;
     }
 
-    @Override
-    protected void pluginInitialize() {
-        insetMask = DEFAULT_INSET_MASK;
+    public static class ListenerConfiguration {
+        public Integer mask;
+        public boolean includeRoundedCorners;
 
-        ViewCompat.setOnApplyWindowInsetsListener(
-            this.cordova.getActivity().findViewById(android.R.id.content), (v, insetProvider) -> {
-                JSONObject result = new JSONObject();
-                
-                try {
-                    float density = this.cordova.getActivity().getResources().getDisplayMetrics().density;
+        public ListenerConfiguration() {
+            mask = null;
+            includeRoundedCorners = DEFAULT_INCLUDE_ROUNDED_CORNERS;
+        }
+    }
 
-                    // Ideally, we'd import this, but it shares the same name as our plugin
-                    androidx.core.graphics.Insets insets = insetProvider.getInsets(insetMask);
+    public static class Listener {
+        private final Context $context;
+        private final CallbackContext $callback;
+        private JSONObject $currentInsets;
+        private final int $mask;
+        private final boolean $includeRoundedCorners;
+        private final UUID $id;
 
-                    double topLeftRadius = 0.0;
-                    double topRightRadius = 0.0;
-                    double botLeftRadius = 0.0;
-                    double botRightRadius = 0.0;
+        public Listener(Context context, CallbackContext callback, ListenerConfiguration config) {
+            $id = UUID.randomUUID();
+            $context = context;
+            $callback = callback;
+            if (config.mask == null) {
+                $mask = DEFAULT_INSET_MASK;
+            }
+            else {
+                $mask = $mapMask(config.mask);
+            }
+            $includeRoundedCorners = config.includeRoundedCorners;
+        }
 
+        public String getID() {
+            return $id.toString();
+        }
+
+        public void onInsetUpdate(WindowInsetsCompat insetProvider) {
+            JSONObject result = new JSONObject();
+
+            try {
+                float density = $context.getResources().getDisplayMetrics().density;
+
+                // Ideally, we'd import this, but it shares the same name as our plugin
+                androidx.core.graphics.Insets insets = insetProvider.getInsets($mask);
+
+                double topLeftRadius = 0.0;
+                double topRightRadius = 0.0;
+                double botLeftRadius = 0.0;
+                double botRightRadius = 0.0;
+
+                if ($includeRoundedCorners && Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
                     WindowInsets sourceInsets = insetProvider.toWindowInsets();
                     if (sourceInsets != null) {
-                        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
                             RoundedCorner topLeft = sourceInsets.getRoundedCorner(RoundedCorner.POSITION_TOP_LEFT);
                             RoundedCorner topRight = sourceInsets.getRoundedCorner(RoundedCorner.POSITION_TOP_RIGHT);
                             RoundedCorner botLeft = sourceInsets.getRoundedCorner(RoundedCorner.POSITION_BOTTOM_LEFT);
@@ -81,124 +111,201 @@ public class Insets extends CordovaPlugin {
 
                             if (topLeft != null) {
                                 int radius = topLeft.getRadius();
-                                topLeftRadius = (double)radius / density;
+                                topLeftRadius = (double) radius / density;
                             }
 
                             if (topRight != null) {
                                 int radius = topRight.getRadius();
-                                topRightRadius = (double)radius / density;
+                                topRightRadius = (double) radius / density;
                             }
 
                             if (botLeft != null) {
                                 int radius = botLeft.getRadius();
-                                botLeftRadius = (double)radius / density;
+                                botLeftRadius = (double) radius / density;
                             }
 
                             if (botRight != null) {
                                 int radius = botRight.getRadius();
-                                botRightRadius = (double)radius / density;
+                                botRightRadius = (double) radius / density;
                             }
+                        }
+                }
+
+                double top = insets.top / density;
+                double right = insets.right / density;
+                double bottom = insets.bottom / density;
+                double left = insets.left / density;
+
+                // Insets do not include rounded corner radius. If an inset is present, it
+                // generally will be big enough to cover the rounded corner. This is a coincidence,
+                // not a designed thing. In either case, we need to determine how much space is
+                // required to cover the rounded corner and take the higher between the inset and
+                // the rounded corner.
+
+                top = Math.max(Math.max(top, topLeftRadius), topRightRadius);
+                bottom = Math.max(Math.max(bottom, botLeftRadius), botRightRadius);
+                left = Math.max(Math.max(left, topLeftRadius), botLeftRadius);
+                right = Math.max(Math.max(right, topRightRadius), botRightRadius);
+
+                result.put("top", top);
+                result.put("right", right);
+                result.put("bottom", bottom);
+                result.put("left", left);
+            }
+            catch (JSONException e) {
+                e.printStackTrace();
+                return;
+            }
+
+            this.$currentInsets = result;
+
+            JSONObject update = new JSONObject();
+            try {
+                update.put("type", "update");
+                update.put("id", this.getID());
+                update.put("data", this.$currentInsets);
+            }
+            catch (JSONException ex) {
+                throw new RuntimeException(ex);
+            }
+
+            PluginResult response = new PluginResult(Status.OK, update);
+            response.setKeepCallback(true);
+            $callback.sendPluginResult(response);
+        }
+
+        private int $mapMask(int webviewMask) {
+            int insetTypeMask = 0;
+
+            if ((webviewMask & WebviewMask.CAPTION_BAR) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.captionBar();
+            }
+
+            if ((webviewMask & WebviewMask.DISPLAY_CUTOUT) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.displayCutout();
+            }
+
+            if ((webviewMask & WebviewMask.IME) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.ime();
+            }
+
+            if ((webviewMask & WebviewMask.MANDATORY_SYSTEM_GESTURES) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.mandatorySystemGestures();
+            }
+
+            if ((webviewMask & WebviewMask.NAVIGATION_BARS) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.navigationBars();
+            }
+
+            if ((webviewMask & WebviewMask.STATUS_BARS) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.statusBars();
+            }
+
+            if ((webviewMask & WebviewMask.SYSTEM_BARS) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.systemBars();
+            }
+
+            if ((webviewMask & WebviewMask.SYSTEM_GESTURES) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.systemGestures();
+            }
+
+            if ((webviewMask & WebviewMask.TAPPABLE_ELEMENT) != 0) {
+                insetTypeMask |= WindowInsetsCompat.Type.tappableElement();
+            }
+
+            return insetTypeMask;
+        }
+    }
+
+    private ArrayList<Listener> $listeners;
+    private HashMap<String, Listener> $listenerMap;
+    private final Object $listenerLock = new Object();
+
+    @Override
+    protected void pluginInitialize() {
+        $listeners = new ArrayList<>();
+        $listenerMap = new HashMap<>();
+        
+        ViewCompat.setOnApplyWindowInsetsListener(
+                this.cordova.getActivity().findViewById(android.R.id.content), (v, insetProvider) -> {
+
+                    synchronized($listenerLock) {
+                        for (Listener listener : $listeners) {
+                            listener.onInsetUpdate(insetProvider);
                         }
                     }
 
-                    double top = insets.top / density;
-                    double right = insets.right / density;
-                    double bottom = insets.bottom / density;
-                    double left = insets.left / density;
-
-                    // Insets do not include rounded corner radius. If an inset is present, it generally will be big enough to cover the rounded corner. This is a coincidence, not a designed thing.
-                    // In either case, we need to determine how much space is required to cover the rounded corner and take the higher betwen the inset and the rounded corner.
-
-                    top = Math.max(Math.max(top, topLeftRadius), topRightRadius);
-                    bottom = Math.max(Math.max(bottom, botLeftRadius), botRightRadius);
-                    left = Math.max(Math.max(left, topLeftRadius), botLeftRadius);
-                    right = Math.max(Math.max(right, topRightRadius), botRightRadius);
-                    
-                    result.put("top", top);
-                    result.put("right", right);
-                    result.put("bottom", bottom);
-                    result.put("left", left);
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                    return insetProvider.CONSUMED; // Stop dispatching to child views
+                    return WindowInsetsCompat.CONSUMED;
                 }
-                this.insets = result;
-                if (listener != null) {
-                    PluginResult presult = new PluginResult(Status.OK, this.insets);
-                    presult.setKeepCallback(true);
-                    listener.sendPluginResult(presult);
-                }
-                return insetProvider.CONSUMED; // Stop dispatching to child views
-            }
         );
+    }
+
+    private void $createNewListener(CallbackContext callback, JSONArray args) {
+        ListenerConfiguration config = new ListenerConfiguration();
+
+        try {
+            JSONObject params = args.getJSONObject(0);
+            if (params.has("mask")) {
+                config.mask = params.getInt("mask");
+            }
+            if (params.has("includeRoundedCorners")) {
+                config.includeRoundedCorners = params.getBoolean("includeRoundedCorners");
+            }
+        }
+        catch (JSONException ex) {
+            ex.printStackTrace();
+            callback.error(ex.getMessage());
+            return;
+        }
+        Listener listener = new Listener(cordova.getActivity(), callback, config);
+        synchronized ($listenerLock) {
+            $listeners.add(listener);
+            $listenerMap.put(listener.getID(), listener);
+        }
+
+        cordova.getActivity().runOnUiThread(() -> {
+            ViewCompat.requestApplyInsets(cordova.getActivity().findViewById(android.R.id.content));
+        });
+
+        JSONObject responseData = new JSONObject();
+        try {
+            responseData.put("type", "init");
+            responseData.put("data", listener.getID());
+        }
+        catch(JSONException e) {
+            throw new RuntimeException("Could not build listener creation response", e);
+        }
+
+        PluginResult result = new PluginResult(Status.OK, responseData);
+        result.setKeepCallback(true);
+        callback.sendPluginResult(result);
+    }
+
+    private void $freeListener(CallbackContext callback, JSONArray args) throws JSONException {
+        String id = args.getString(0);
+
+        synchronized ($listenerLock) {
+            Listener listener = $listenerMap.getOrDefault(id, null);
+            $listenerMap.remove(id);
+            if (listener != null) {
+                $listeners.remove(listener);
+            }
+        }
+
+        callback.success();
     }
 
     @Override
     public boolean execute(String action, JSONArray args, CallbackContext callback) throws JSONException, NumberFormatException {
-        if (action.equals("setListener")) {
-            listener = callback;
-            cordova.getActivity().runOnUiThread(() -> {
-                if (this.insets == null) {
-                    // Trigger a inset dispatch (will eventually jump to the ApplyWindowInsetsListener above)
-                    ViewCompat.requestApplyInsets(cordova.getActivity().findViewById(android.R.id.content));
-                }
-                else {
-                    PluginResult presult = new PluginResult(Status.OK, this.insets);
-                    presult.setKeepCallback(true);
-                    listener.sendPluginResult(presult);
-                }
-            });
+        if (action.equals("create")) {
+            $createNewListener(callback, args);
             return true;
         }
-        else if (action.equals("setMask")) {
-            cordova.getActivity().runOnUiThread(() -> {
-                int mask = args.getInt(0);
-                ViewCompat.requestApplyInsets(cordova.getActivity().findViewById(android.R.id.content));
-                // TODO: Not using callback context and wanting to return the insets, but perhaps we should
-                // just make it return void and let the listener propagate the change?
-            });
+        else if (action.equals("delete")) {
+            $freeListener(callback, args);
             return true;
         }
+
         return false;
-    }
-
-    private int mapMask(int webviewMask) {
-        int insetTypeMask = 0;
-
-        if ((webviewMask & WebviewMask.CAPTION_BAR) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.captionBar();
-        }
-
-        if ((webviewMask & WebviewMask.DISPLAY_CUTOUT) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.displayCutout();
-        }
-
-        if ((webviewMask & WebviewMask.IME) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.ime();
-        }
-
-        if ((webviewMask & WebviewMask.MANDATORY_SYSTEM_GESTURES) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.mandatorySystemGestures();
-        }
-
-        if ((webviewMask & WebviewMask.NAVIGATION_BARS) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.navigationBars();
-        }
-
-        if ((webviewMask & WebviewMask.STATUS_BARS) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.statusBars();
-        }
-
-        if ((webviewMask & WebviewMask.SYSTEM_BARS) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.systemBars();
-        }
-
-        if ((webviewMask & WebviewMask.SYSTEM_GESTURES) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.systemGestures();
-        }
-
-        if ((webviewMask & WebviewMask.TAPPABLE_ELEMENT) != 0) {
-            insetTypeMask |= WindowInsetsCompat.Type.tappableElement();
-        }
     }
 }